+++ aerokit/aero/model1D.py
@@ -6,7 +6,7 @@
 import numpy     as np
 #import ShockWave as sw
 #import IterativeSolve
-from ..common   import defaultgas as defg
+from aerokit.common import defaultgas as defg
 from aerokit.aero import Isentropic as Is
 
 # -- class --
@@ -21,14 +21,18 @@ class state():
 		u
 		p
 	"""
-	def __init__(self, rho, u, p, gamma=defg._gamma):
-	 	self._gamma = gamma
-	 	self.rho    = rho
-	 	self.u      = u
-	 	self.p      = p
+	def __init__(self, rho = None, u = None, p = None, gamma=defg._gamma):
+		self._gamma = gamma
+		self.rho    = rho
+		self.u      = u
+		self.p      = p
 
 	def __repr__(self):
 		return "state (rho, u, p) : (%s, %s, %s)" % (self.rho, self.u, self.p)
+
+	@property
+	def size(self): 
+		return self.rho.size if isinstance(self.rho, np.ndarray) else 1
 
 	def copy(self):
 		return state(self.rho, self.u, self.p, self._gamma)
@@ -56,7 +60,7 @@ class state():
 		M   = Is.Mach_PtPs(pt/p, self._gamma)
 		rts = rtt/Is.TtTs_Mach(M, self._gamma)
 		self.__init__(rho=p/rts, u=M*np.sqrt(self._gamma*rts), p=p)
-		
+
 	def asound(self):
 		"""returns speed of sound"""
 		return np.sqrt(self._gamma*self.p/self.rho)